Fixed a nasty integer overflow bug P is now a full-fledged transpose matrix Fixed a math problem with L New func to round whole matrix to precision Det potentially still bugged in terms of sign

+++ src/commonMain/kotlin/fraction/IntList.kt
@@ -10,12 +10,12 @@ class IntList(vararg values: Int) {
     var memberCount = 0
     internal var root: IntMember? = null
 
-    private var prod: Int = 0
+    private var prod: Long = 0
     private var prodReady = false
 
     inner class IntMember(var value: Int, var next: IntMember? = null)
 
-    fun product(): Int {
+    fun product(): Long {
         if (prodReady)
             return prod
